refactor(test): simplify framework HTTP fixture helpers
- Move bearer and CSRF request helpers into the HTTP handler fixture
- Keep CSRF configuration private behind AddCsrfToken
- Allow missing RDBMS and ClickHouse test migrations during fixture setup
