--- internal/test/fixture/http/bearer.go
@@ -1,7 +1,0 @@
-package http
-
-import "net/http"
-
-func (handler *Handler) AddBearerToken(req *http.Request, token string) {
-	req.Header.Set("Authorization", token)
-}
