+++ internal/test/fixture/http/handler.go
@@ -55,3 +55,23 @@ func New(dependencies Dependencies, routeParams RouteParams) *Handler {
 func (handler *Handler) ServeHTTP(w http.ResponseWriter, req *http.Request) {
 	handler.engine.ServeHTTP(w, req)
 }
+
+func (handler *Handler) AddBearerToken(req *http.Request, token string) {
+	req.Header.Set("Authorization", token)
+}
+
+func (handler *Handler) AddCsrfToken(req *http.Request) {
+	config := handler.csrfConfig
+	cookie := &http.Cookie{
+		Name:     config.Cookie.Name,
+		Value:    "1234567890",
+		Path:     config.Cookie.Path,
+		Domain:   config.Cookie.Domain,
+		MaxAge:   config.Cookie.MaxAge,
+		Secure:   config.Cookie.Secure,
+		HttpOnly: config.Cookie.HttpOnly,
+		SameSite: config.Cookie.SameSite,
+	}
+	req.AddCookie(cookie)
+	req.Header.Set(config.Header, "1234567890")
+}
